requirements.txt
Django==3.2.4
django-datatables-view==1.19.1
django-extensions==3.0.9
django-filter
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.17.1
Jinja2==3.0.1
Pillow==8.1.2
PyJWT==2.0.1
requests==2.25.1
Unidecode==1.2.0
urllib3==1.26.5
python-dotenv==0.19.2
humanize
badi-utils